Move play_audio function to audio.music module
- Create new audio/music.py module for audio playback functionality
- Move play_audio() from tone_detect_test.py to music.py
- Update imports in tone_detect_test.py to use the new module
- Maintains all existing functionality with better separation of concerns
